requirements.txt
numpy==1.24.1
pillow==8.3.1
scipy==1.10.0
requests==2.26.0
tqdm==4.62.2
ninja==1.11.1
matplotlib==3.4.2
imageio==2.9.0
glfw==2.2.0
pyopengl==3.1.5
imageio-ffmpeg==0.4.3
open-clip-torch==2.10.1
webdataset==0.2.31
setuptools==58.0.4
psutil==5.4.7
dill==0.3.6
tensorboard==2.11.2
click==8.1.3
timm==0.6.12
